+++ src/Renci.SshNet/SshMessageFactory.cs
@@ -16,7 +16,7 @@ namespace Renci.SshNet
         private readonly bool[] _activatedMessagesById;
         private readonly object _lock = new object();
 
-        private static readonly MessageMetadata[] AllMessages = new MessageMetadata[]
+        internal static readonly MessageMetadata[] AllMessages = new MessageMetadata[]
             {
                 new MessageMetadata<KeyExchangeInitMessage>(0, "SSH_MSG_KEXINIT", 20),
                 new MessageMetadata<NewKeysMessage>(1, "SSH_MSG_NEWKEYS", 21),
@@ -263,7 +263,7 @@ namespace Renci.SshNet
                                                  currentEnabledForMessageName));
         }
 
-        private abstract class MessageMetadata
+        internal abstract class MessageMetadata
         {
             protected MessageMetadata(byte id, string name, byte number)
             {
