--- test/Renci.SshNet.Tests/Common/HttpRequest.cs
@@ -1,17 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Renci.SshNet.Tests.Common
-{
-    public class HttpRequest
-    {
-        public HttpRequest()
-        {
-            Headers = new List<string>();
-            MessageBody = new byte[0];
-        }
-
-        public string RequestLine { get; set; }
-        public IList<string> Headers { get; set; }
-        public byte[] MessageBody { get; set; }
-    }
-}
